Do not implement SharedListenerAggregate
We're considering removing this in v3 of the EventManager, and it's essentially
unnecessary anyways, as the "aggregate" only has a single method. This patch
modifies the `Module` class to attach using a closure (more below), and removes
the methods in `RestParametersListener` that implement the
`SharedListenerAggregate` functionality.

In order to reduce performance impact, the closure passed to the shared event
manager wraps the call for retrieving the listener, ensuring that the
performance hit of retrieval is delayed until it is actually used. We then
return the value of invoking the listener method directly.